+++ Assets/com.liltoon.pcss-extension-1.5.8/Editor/AvatarSelectorMenu.cs
@@ -1,0 +1,3 @@
+#if MODULAR_AVATAR_AVAILABLE
+using nadena.dev.modular_avatar.core;
+#endif